partial api breach - move _CompatProperty to only python nodes - they are only used there
