+++ src/_pytest/python.py
@@ -11,6 +11,7 @@ import sys
 import warnings
 from textwrap import dedent
 
+import attr
 import py
 import six
 
@@ -125,10 +126,10 @@ def pytest_generate_tests(metafunc):
     # those alternative spellings are common - raise a specific error to alert
     # the user
     alt_spellings = ["parameterize", "parametrise", "parameterise"]
-    for attr in alt_spellings:
-        if hasattr(metafunc.function, attr):
+    for spelling in alt_spellings:
+        if hasattr(metafunc.function, spelling):
             msg = "{0} has '{1}' mark, spelling should be 'parametrize'"
-            fail(msg.format(metafunc.function.__name__, attr), pytrace=False)
+            fail(msg.format(metafunc.function.__name__, spelling), pytrace=False)
     for marker in metafunc.definition.iter_markers(name="parametrize"):
         metafunc.parametrize(*marker.args, **marker.kwargs)
 
@@ -235,7 +236,43 @@ class PyobjContext(object):
     instance = pyobj_property("Instance")
 
 
-class PyobjMixin(PyobjContext):
+@attr.s
+class _CompatProperty(object):
+    name = attr.ib()
+
+    def __get__(self, obj, owner):
+        if obj is None:
+            return self
+
+        from _pytest.deprecated import COMPAT_PROPERTY
+
+        warnings.warn(
+            COMPAT_PROPERTY.format(name=self.name, owner=owner.__name__), stacklevel=2
+        )
+        return getattr(__import__("pytest"), self.name)
+
+
+class _CompatPropertyMixin(object):
+    Module = _CompatProperty("Module")
+    Class = _CompatProperty("Class")
+    Instance = _CompatProperty("Instance")
+    Function = _CompatProperty("Function")
+    File = _CompatProperty("File")
+    Item = _CompatProperty("Item")
+
+    def _getcustomclass(self, name):
+        maybe_compatprop = getattr(type(self), name)
+        if isinstance(maybe_compatprop, _CompatProperty):
+            return getattr(__import__("pytest"), name)
+        else:
+            from _pytest.deprecated import CUSTOM_CLASS
+
+            cls = getattr(self, name)
+            self.warn(CUSTOM_CLASS.format(name=name, type_name=type(self).__name__))
+        return cls
+
+
+class PyobjMixin(_CompatPropertyMixin, PyobjContext):
     _ALLOW_MARKERS = True
 
     def __init__(self, *k, **kw):
